Format code

+++ src/main.rs
@@ -397,8 +397,8 @@ fn main() {
         let center = (32, 32);
         for y in 0..64 {
             for x in 0..64 {
-                let dx = x as i32 - center.0 as i32;
-                let dy = y as i32 - center.1 as i32;
+                let dx = x as i32 - center.0;
+                let dy = y as i32 - center.1;
                 if dx*dx + dy*dy < 15*15 {
                     img.put_pixel(x, y, image::Luma([255u8]));
                 }
